requirements.txt
asgiref==3.8.1
bcrypt==4.0.1
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.68.1
h11==0.14.0
motor==2.5.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.19
pymongo==3.13.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
starlette==0.14.2
typing_extensions==4.12.2
uvicorn==0.15.0
